consolidate shared layout styles
