+++ src/components/FormSection/FormSection.tsx
@@ -1,5 +1,6 @@
 import { styled } from "next-yak";
 import { theme } from "@/styles/theme.yak";
+import { sharedSectionHeadingStyles } from "@/styles/commonStyles";
 import type { PropsWithChildren } from "react";
 
 const StyledSection = styled.section`
@@ -10,12 +11,9 @@ const StyledSection = styled.section`
     padding-bottom: 2rem;
     border-bottom: 1px solid ${theme.colors.border.base};
   }
-  & h2,
-  & h3 {
-  }
+
   & h2 {
-    font-size: 1.5rem;
-    font-weight: 600;
+    ${sharedSectionHeadingStyles};
   }
 `;
